pyproject.toml
[project]
name = "echo"
version = "0.1.0"
description = "Collective Intelligence Infrastructure for AI Agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "qdrant-client>=1.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "boto3>=1.35.0",
    "wandb>=0.18.0",
    "datasets>=3.0.0",
    "tqdm>=4.66.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
